edited script/task for adding dummy products

+++ auctions/management/commands/add_dummy_products.py
@@ -30,7 +30,7 @@ class Command(BaseCommand):
 
         for listing in products:
             end_time = timezone.now() + datetime.timedelta(
-                minutes=random.randint(5, 30))
+                minutes=random.randint(10, 14400))
             listing = Listing(title=listing['title'], starting_price=listing['price'], current_price=listing['price'],
                               description=listing['description'], image_url=listing['image'], category=listing['category'], owner=random.choice(users), end_time=end_time)
             listing.save()
@@ -40,4 +40,5 @@ class Command(BaseCommand):
             closed=True).order_by('date_added').all()[:20]
 
         for listing in old_products:
-            listing.delete()
+            if listing.owner in users:
+                listing.delete()
